+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-05-04 21:29:15
-// Commit: style(ui): harmonize light and dark theme palette for ui
+// Updated at: 2026-05-05 16:57:49
+// Commit: refactor(ui): optimize data flow between domain and data layers
 
 class uiComponentState {
-  final String commitRef = '861';
-  final String lastMessage = 'style(ui): harmonize light and dark theme palette for ui';
-  final DateTime timestamp = DateTime.parse('2026-05-04T21:29:15+05:00');
+  final String commitRef = '895';
+  final String lastMessage = 'refactor(ui): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-05-05T16:57:49+05:00');
 }